org.elixir_lang.annonator -> org.elixir_lang.annotator
Fix typo in package name.
