+++ src/org/elixir_lang/annotator/Atom.java
@@ -1,6 +1,5 @@
-package org.elixir_lang.annonator;
+package org.elixir_lang.annotator;
 
-import com.intellij.lang.ASTNode;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.openapi.editor.colors.EditorColorsManager;
@@ -10,19 +9,15 @@ import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiRecursiveElementVisitor;
+import com.intellij.psi.impl.source.tree.LeafPsiElement;
 import org.elixir_lang.ElixirSyntaxHighlighter;
-import org.elixir_lang.psi.ElixirMapArguments;
-import org.elixir_lang.psi.ElixirMapOperation;
-import org.elixir_lang.psi.ElixirStructOperation;
+import org.elixir_lang.psi.ElixirKeywordKey;
 import org.jetbrains.annotations.NotNull;
 
-import static org.elixir_lang.ElixirSyntaxHighlighter.BRACES_TOKEN_SET;
-
-
 /**
- * Annotates maps and structs
+ * Annotates things that act like Atom as Atom
  */
-public class Map implements Annotator, DumbAware {
+public class Atom implements Annotator, DumbAware {
     /*
      * Public Instance Methods
      */
@@ -46,10 +41,8 @@ public class Map implements Annotator, DumbAware {
 
                    @Override
                    public void visitElement(PsiElement element) {
-                       if (element instanceof ElixirMapOperation) {
-                         visitMapOperation((ElixirMapOperation) element);
-                       } else if (element instanceof ElixirStructOperation) {
-                           visitStructOperation((ElixirStructOperation) element);
+                       if (element instanceof ElixirKeywordKey) {
+                         visitKeywordKey((ElixirKeywordKey) element);
                        }
                    }
 
@@ -57,16 +50,19 @@ public class Map implements Annotator, DumbAware {
                     * Private Instance Methods
                     */
 
-                   private void visitMapOperation(ElixirMapOperation mapOperation) {
-                       highlight(mapOperation.getMapPrefixOperator(), holder, ElixirSyntaxHighlighter.MAP);
-                       highlight(mapOperation.getMapArguments(), holder, ElixirSyntaxHighlighter.MAP);
-                   }
+                   private void visitKeywordKey(ElixirKeywordKey keywordKey) {
+                       PsiElement child = keywordKey.getFirstChild();
 
-                   private void visitStructOperation(ElixirStructOperation structOperation) {
-                       highlight(structOperation.getMapPrefixOperator(), holder, ElixirSyntaxHighlighter.STRUCT);
-                       /* DO NOT highlight mapExpression.  It will be highlighted as either an alias or module
-                          attribute, which are more specific and useful. */
-                       highlight(structOperation.getMapArguments(), holder, ElixirSyntaxHighlighter.STRUCT);
+                       // a normal, non-quoted keyword key
+                       if (child instanceof LeafPsiElement) {
+                           TextRange keywordKeyTextRange = keywordKey.getTextRange();
+                           // highlight the `:` as part of the pseudo-atom
+                           TextRange atomTextRange = new TextRange(
+                                   keywordKeyTextRange.getStartOffset(),
+                                   keywordKeyTextRange.getEndOffset() + 1
+                           );
+                           highlight(atomTextRange, holder, ElixirSyntaxHighlighter.ATOM);
+                       }
                    }
                }
        );
@@ -75,22 +71,6 @@ public class Map implements Annotator, DumbAware {
     /*
      * Private Instance Methods
      */
-
-    private void highlight(@NotNull final ElixirMapArguments mapArguments,
-                           @NotNull AnnotationHolder annotationHolder,
-                           @NotNull final TextAttributesKey textAttributesKey) {
-        ASTNode[] braces = mapArguments.getNode().getChildren(BRACES_TOKEN_SET);
-
-        for (ASTNode brace : braces) {
-            highlight(brace.getTextRange(), annotationHolder, textAttributesKey);
-        }
-    }
-
-    private void highlight(@NotNull final PsiElement element,
-                           @NotNull AnnotationHolder annotationHolder,
-                           @NotNull final TextAttributesKey textAttributesKey) {
-        highlight(element.getTextRange(), annotationHolder, textAttributesKey);
-    }
 
     /**
      * Highlights `textRange` with the given `textAttributesKey`.
